+++ test/build_gfonts.dart
@@ -128,8 +128,12 @@ void main(List<String> args) async {
   }
 
   for (final entry in <String, String>{
+    'CupertinoIcons':
+        'https://github.com/flutter/packages/blob/master/third_party/packages/cupertino_icons/assets/CupertinoIcons.ttf',
     'MaterialIcons':
         'https://fonts.gstatic.com/s/materialicons/v98/flUhRq6tzZclQEJ-Vdg-IuiaDsNZ.ttf',
+    'NotoColorEmoji':
+        'https://github.com/googlefonts/noto-emoji/raw/main/fonts/NotoColorEmoji.ttf',
   }.entries) {
     output.writeln('');
     output.writeln('/// ${entry.key}');